requirements.txt
apache-airflow==3.1.4
pandas==2.3.3
pymongo==4.15.5
apache-airflow-providers-mongo==5.3.0